depend bigcouch package install on ::resolvconf
before, it depended on site_config::resolvconf, which didn't work
because site_config::resolvconf included ::resolvconf, and
therefore doesn't *contain* it.

from http://puppetlabs.com/blog/class-containment-puppet:
"This is because classes never contain the classes they include. So even
though classa and classb are ordered, the classes they include are not."

Change-Id: Icf85bf68f007bcc78edbd6d863c45eaf7a967e82